requirements.txt
python-dotenv
openai
gradio
pypdf
requests
beautifulsoup4
pydub

